deps/v8/test/webkit/fast/js/kde/object_prototype-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sub.x is 1
PASS sub.y is 2
PASS sub.hasOwnProperty('x') is false
PASS sub.hasOwnProperty('y') is true
PASS sub.x is 6
PASS sub.hasOwnProperty('x') is true
PASS sub.y is undefined
PASS sub.hasOwnProperty('y') is false
PASS obj.hasClass1 is true
PASS obj.hasClass2 is true
PASS obj.hasClass3 is true
PASS Class1.prototype.isPrototypeOf(obj) is true
PASS Class2.prototype.isPrototypeOf(obj) is true
PASS Class3.prototype.isPrototypeOf(obj) is true
PASS obj.isPrototypeOf(Class1.prototype) is false
PASS obj.isPrototypeOf(Class2.prototype) is false
PASS obj.isPrototypeOf(Class3.prototype) is false
PASS Class1.prototype.isPrototypeOf(Class2.prototype) is true
PASS Class2.prototype.isPrototypeOf(Class1.prototype) is false
PASS Class1.prototype.isPrototypeOf(Class3.prototype) is true
PASS Class3.prototype.isPrototypeOf(Class1.prototype) is false
PASS Class2.prototype.isPrototypeOf(Class3.prototype) is true
PASS Class3.prototype.isPrototypeOf(Class2.prototype) is false
PASS Class1.prototype.prototype is undefined.
PASS myfunc.length is 3
PASS myfunc.someproperty is 4
PASS myfunc.propertyIsEnumerable('length') is false
PASS myfunc.propertyIsEnumerable('someproperty') is true
PASS checkEnumerable(myfunc,'length') is false
PASS checkEnumerable(myfunc,'someproperty') is true
PASS successfullyParsed is true

TEST COMPLETE
